+++ src/main/kotlin/com/tul/shoppingcart/validations/RequestValidator.kt
@@ -3,6 +3,8 @@ package com.tul.shoppingcart.validations
 import com.tul.shoppingcart.dto.ProductDTO
 import com.tul.shoppingcart.exceptions.BadRequestException
 import com.tul.shoppingcart.exceptions.ErrorCode.*
+import java.util.*
+import java.util.stream.Stream
 
 object RequestValidator {
 
@@ -13,7 +15,15 @@ object RequestValidator {
         request.sku ?: throw BadRequestException(SKU_NOT_FOUND)
     }
 
-    fun validateField(id: String?) {
+    fun validateId(id: String?) {
         id ?: throw BadRequestException(ID_NOT_FOUND)
+    }
+
+    fun validateUpdateRequest(id: String?, request: ProductDTO?) {
+        validateId(id)
+        Stream.of(request!!.description, request.sku, request.name, request.price, request.hasDiscount)
+            .allMatch(Objects::isNull).let { result ->
+                if (result) throw BadRequestException(NOT_NULL_EXPECTED)
+            }
     }
 }